# machine/index: make index a tree for fast traverse ops