Currently index is not suitable for get operations - in order to list all files under specific directory, one would need to get all index keys, and match them with path prefix of the specific directory. In case of big mounts, this operation is not feasible (e.g. koding repo can have around 330k files when built).

This PR reworks the underlying data structure to a tree.

Coverage: 98.5%

As a side-goal the index size dropped by 13.25% (due to the fact paths prefixes are stored only once).

Before:

```
$ go test -run TestRepo -repo /Users/rjeczalik/src/github.com/koding/koding -v
=== RUN   TestRepo
--- PASS: TestRepo (29.45s)
        benchmark_test.go:41: Index build time: 26.226116212s
        benchmark_test.go:42: Index file count: 409715
        benchmark_test.go:43: Index size: 10.1013 MiB
PASS
ok      koding/klient/machine/index     29.500s
```

After:

```
$ go test -run TestRepo -repo /Users/rjeczalik/src/github.com/koding/koding -v
=== RUN   TestRepo
--- PASS: TestRepo (28.87s)
        benchmark_test.go:31: Index build time: 26.516737166s
        benchmark_test.go:32: Index file count: 409715
        benchmark_test.go:33: Index size: 8.7630 MiB
PASS
ok      koding/klient/machine/index     28.925s
```