@rjeczalik 2017-02-03
> and put some comments to public methods?

Yup, wanted to do that separately, this PR already has >300 loc.