@rjeczalik 2017-02-03
I'm going to send inline doc update in a separate PR.